more frontmatter updates
