+++ notebooks/week14/optimization_with_JuMP.jl
@@ -8,9 +8,10 @@
 #> section = 11
 #> order = 11
 #> title = "Solving inverse problems"
+#> layout = "layout.jlhtml"
 #> youtube_id = "nm86_hDwYTU"
-#> tags = ["lecture", "module3"]
 #> description = ""
+#> tags = ["lecture", "module3"]
 
 using Markdown
 using InteractiveUtils
