+++ src/week2/transforming_images.jl
@@ -3,14 +3,15 @@
 
 #> [frontmatter]
 #> chapter = 1
-#> video = "https://www.youtube.com/watch?v=KyBXJV1zFlo"
-#> image = "https://user-images.githubusercontent.com/6933510/136196599-c6ae60f0-9269-4134-bb0d-5bcab928bd2b.gif"
-#> section = 7
-#> order = 7
-#> title = "Dynamic Programming"
-#> youtube_id = "KyBXJV1zFlo"
-#> tags = ["lecture", "module1"]
+#> video = "https://www.youtube.com/watch?v=uZYVjDDZW9A"
+#> image = "https://user-images.githubusercontent.com/6933510/136196626-194e81c9-00f7-49f6-90c3-09945723b6a3.png"
+#> section = 4
+#> order = 4
+#> title = "Transformations with Images"
+#> layout = "layout.jlhtml"
+#> youtube_id = "uZYVjDDZW9A"
 #> description = ""
+#> tags = ["lecture", "module1"]
 
 using Markdown
 using InteractiveUtils
@@ -25,266 +26,469 @@ macro bind(def, element)
     end
 end
 
-# ╔═╡ 71b53b98-8038-11eb-0ea5-d953294e9f35
+# ╔═╡ 86f770fe-74a1-11eb-01f7-5b3ecf057124
 begin
-	import ImageIO
-	using Plots, PlutoUI, Colors, Images
+	using PlutoUI 
+	using Colors, ColorVectorSpace, ImageShow, FileIO, ImageIO
+	using Unitful 
+	using ImageFiltering
+	using OffsetArrays
+	using Plots
+
+	# Small patch to make images look more crisp:
+	# https://github.com/JuliaImages/ImageShow.jl/pull/50
+	Base.showable(::MIME"text/html", ::AbstractMatrix{<:Colorant}) = false
 end
 
-# ╔═╡ a84fdba4-80db-11eb-13dc-3f440653b2b9
+# ╔═╡ 8d389d80-74a1-11eb-3452-f38eff03483b
+PlutoUI.TableOfContents(aside=true)
+
+# ╔═╡ 9f1a72da-7532-11eb-079c-b7baccc6614a
 md"""
-## Intro to Dynamic Programming 
+#### Intializing packages
+
+_When running this notebook for the first time, this could take up to 15 minutes. Hang in there!_
 """
 
-# ╔═╡ 938107f0-80ee-11eb-18cf-775802c43c2f
+# ╔═╡ 4d332c7e-74f8-11eb-1f49-a518246d1db8
 md"""
-What is dynamic progamming? The word "programming" here is a rather archaic word (but still in  use) for an **optimization problem**, as used, for example, in the phrase 
-"linear programming."  Probably the word "programming" should be abandoned in this context, but no doubt it is too late.
+# Announcement: Lectures will be nearly an hour
 """
 
-# ╔═╡ eb043a90-8102-11eb-3b78-d590a23c83f4
+# ╔═╡ f7689472-74a8-11eb-32a1-8379ae5c88e1
+rotabook = PlutoUI.Resource("https://i.gr-assets.com/images/S/compressed.photo.goodreads.com/books/1348902666l/1646354.jpg")
+
+# ╔═╡ 0f2f9004-74a8-11eb-01a2-973dbe80f166
 md"""
-### Summing over paths problem
+##  **Never run overtime** (a microcentury with UnitFul)
+
+Running overtime is the one unforgivable error a lecturer can make.
+After fifty minutes (one microcentury as von Neumann used to say)
+everybody's attention will turn elsewhere even if we are trying to prove
+the Riemann hypothesis. One minute overtime can destroy the best of
+lectures. (from "Indiscrete Thoughts" by Rota, Chpt 18, 10 Lessons I Wish I Had Been Taught)
 """
 
-# ╔═╡ 5994117c-8102-11eb-1b05-671b7cf87a7e
-md"""
-Let's start by looking at the following problem. 
-Let's create a random matrix and follow paths on it.
-The paths start at one of the square on the top, and can only go downwards, either South-East, South, or South-West.
+# ╔═╡ 962143a8-74a7-11eb-26c3-c10548f326ee
+century = 100u"yr" #  a u"yr" is a special kind of string denoting a unit of a year
 
-We will *add up* the numbers visited along each path. Our goal is to find the path that has the *smallest* sum. So this is indeed an optimization problem: we want to **minimize** the sum along these particular paths.
+# ╔═╡ c2964c80-74f8-11eb-3a74-b1bdd9e4ae02
+century * 2
+
+# ╔═╡ caf488d8-74f8-11eb-0075-0586d66c23c1
+century/200
+
+# ╔═╡ 02dd4a02-74f9-11eb-3d1e-53d83cee8062
+century^2
+
+# ╔═╡ 10ef13d2-74f9-11eb-2849-fb9f83db6ae9
+g = 9.8u"m"/u"s"^2
+
+# ╔═╡ b76a56f4-74a9-11eb-1739-fbfc5e4958e8
+
+uconvert(u"minute", century * 1e-6 ) # convert into minutes the value of a microcentury
+
+
+# ╔═╡ 77fbf18a-74f9-11eb-1d9e-3f9d2097388f
+PotentialEnergy = (10u"kg") * g * (50u"m")
+
+# ╔═╡ bcb69db6-74f9-11eb-100a-29d1d23963ab
+uconvert( u"J",PotentialEnergy)
+
+# ╔═╡ fc70c4d2-74f8-11eb-33f5-539c278ed6b6
+md"""
+Adding units to numbers **just works** in Julia, and furthermore, does not slow down execution.  We are sneaking in an example of the power of generic programming and Julia's type system, some of the underlying technology that makes us love working with Julia.  More on this later in the book.  Meanwhile if this helps you do your problem sets in some other class, go for it.
 """
 
-# ╔═╡ b4558306-804a-11eb-2719-5fd37c6fa281
+# ╔═╡ 2f7cde78-74a2-11eb-1e2f-81b5b2465819
 md"""
-n = $(@bind n Slider(2:12, show_value = true, default=8))
+# Reminder
+
+**Try your own pictures everywhere!**
 """
 
-# ╔═╡ bc631086-804a-11eb-216e-c955e2115f55
-M = rand( 0:9, n, n)
-
-# ╔═╡ 4e4d333e-8102-11eb-0ba1-0f0183d0d3c2
+# ╔═╡ e099815e-74a1-11eb-1541-033f6abe9f8e
 md"""
-One way to solve this problem is the naive algorithm where we enumerate *all* the paths, calculate the sum for each, and take the minimum.
-However, as the matrix gets larger the total number of paths grows *exponentially*.
+# Transforming Images
 """
 
-# ╔═╡ 0f0e7456-8104-11eb-1d90-e9f0009e8789
+# ╔═╡ e82a4dd8-74b0-11eb-1108-6b09e67a80c1
 md"""
-[Possible research problem: Investigate the statistics of the sums over all possible paths.]
+## 2.1. Downsampling / Upsampling
 """
 
-# ╔═╡ 4f969032-80e9-11eb-1ada-d1aa64960967
+# ╔═╡ 39552b7a-74fb-11eb-04e0-3981ada52c92
 md"""
-## Fixing a single point on a path
+How can we pixelate a corgi? Found this cute picture online, but we'll pixelate
+a real corgi.
 """
 
-# ╔═╡ 28f18aa2-8104-11eb-0c01-dbd14c760ecf
+# ╔═╡ 14f2b85e-74ad-11eb-2682-d9de646aedf3
+pixelated_corgi = load(download("https://i.redd.it/99lhfbnwpgd31.png"))
+
+# ╔═╡ 516e73e2-74fb-11eb-213e-9dbd9472e0db
+philip =  load(download("https://user-images.githubusercontent.com/6933510/107239146-dcc3fd00-6a28-11eb-8c7b-41aaf6618935.png"))
+
+# ╔═╡ b5d0ef90-74fb-11eb-3126-792f954c7be7
+@bind r Slider(1:40, show_value=true, default=40)
+
+# ╔═╡ 754c3704-74fb-11eb-1199-2b9798d7251f
+downsample_philip = philip[1:r:end, 1:r:end]
+
+# ╔═╡ 9eb917ba-74fb-11eb-0527-15e981ce9c6a
+upsample_philip = kron(downsample_philip, fill(1,r,r))
+
+# ╔═╡ 486d3022-74ff-11eb-1865-e15436bd9aad
 md"""
-Let's fix a given point $(i, j)$ and focus only on all those paths that pass through $(i, j)$.
+  Note the use of kron and fill. See [Wikipedia Kron](https://en.wikipedia.org/wiki/Kronecker_product)
 """
 
-# ╔═╡ 37ebfa3e-80e5-11eb-166c-4ff3471ab12d
+# ╔═╡ b9da7332-74ff-11eb-241b-fb87e77d646a
 md"""
-i= $(@bind fixi Scrubbable(1:n))
-j= $(@bind fixj Scrubbable(1:n))
+Exercise: Use the nose selection tool from Section 1.1 to pixelate a rectangle of an image.  Warning: you'll have to worry about sizes if not exact multiples.
 """
 
-# ╔═╡ 4d81a6f4-8104-11eb-1f06-5bb7a56c8406
+# ╔═╡ 339ccfca-74b1-11eb-0c35-774da6b189ed
 md"""
-Suppose we fix the point on the penultimate row (last but one). When we look at the paths below the fixed value, we're doing the same calculation over and over again. It doesn't seem sensible to keep re-doing these calculations. The same holds as we move the fixed point further upwards.
-
-So instead of calculating by working "forwards", for each box we look at the minimum below it.
+## 2.2 Linear Combinations (Combining Images) 
 """
 
-# ╔═╡ d9265982-80ed-11eb-3a5f-27712a23506b
+# ╔═╡ 8711c698-7500-11eb-2505-d35a4de169b4
 md"""
-## The idea of *overlapping subproblems*
+One big idea in mathematics is the [linear combination](https://en.wikipedia.org/wiki/Linear_combination).
+The idea combines 
+- scaling an object 
+- combining two or more objects
+by combining scaled versions of multiple objects.
 """
 
-# ╔═╡ ba4acb08-8104-11eb-1771-15bc5d8076fd
+# ╔═╡ 84350cb8-7501-11eb-095e-8f1a7e015f25
 md"""
-The key point in this problem is that there are *overlapping subproblems*: there are calculations that we don't need to repeat. 
-
-The idea of dynamic programming is to remember the solution of those subproblems to get an exponential speed-up in the calculation speed.
+Let's scale some corgis.
 """
 
-# ╔═╡ 163bf8fe-80d0-11eb-2066-75439a533513
+# ╔═╡ 91a1bca4-74aa-11eb-3917-1dfd73d0ad9c
+corgis = load(download("https://user-images.githubusercontent.com/6933510/108605549-fb28e180-73b4-11eb-8520-7e29db0cc965.png"))
+
+# ╔═╡ 8e698bdc-7501-11eb-1d2e-c336ccbde0b0
+@bind c Slider(0:.1:3, show_value=true, default=1)
+
+# ╔═╡ ab2bc924-7501-11eb-03ba-8dfc1ffe3f36
+c .* corgis  # scaling the corgis changes intensity
+
+# ╔═╡ e11d6300-7501-11eb-239a-135596309d20
+md"""
+ You might wonder about the **dot times** or **pointwise times**. You can
+delete the dot, but it is recommended for clarity
+and performance.  The dot emphasizes that the multiplication by c is happening
+pixel by pixel or that the scalar is being "broadcast" to every pixel.
+"""
+
+# ╔═╡ 9a66c07e-7503-11eb-3127-7fce91b3a24a
+md"""
+Scaling too far saturates the image.  (Any r,g,b ≥ 1, saturates at 1.)
+"""
+
+# ╔═╡ 47d40406-7502-11eb-2f43-cd5c848f25a6
+md"""
+We need another image.  We could grab one from somewhere or we can just transform the one we have.  Let's do the latter and turn the corgis upsidedown.
+"""
+
+# ╔═╡ 9ce0b980-74aa-11eb-0678-01209451fb65
+upsidedown_corgis = corgis[ end:-1:1 , :]
+
+# ╔═╡ 68821bf4-7502-11eb-0d3c-03d7a00fdba4
+md"""
+Now let's scaled version of the two images to see what that does.
+"""
+
+# ╔═╡ 447e7c9e-74b1-11eb-27ea-71aa4338b11a
+(.5 * upsidedown_corgis .+ .5 * corgis) 
+
+# ╔═╡ c9dff6f4-7503-11eb-2715-0bf9d3ece9e1
+md"""
+### Convex Combinations
+"""
+
+# ╔═╡ d834103c-7503-11eb-1a94-1fbad43801ff
+md"""
+If all the coefficients are positive and add to 1, we say we have a **convex combination**.  Let's take α and (1-α) as the two coefficients adding to 1, and
+scale the two corgi pictures with different α's, thereby giving different weights to the rightside-up and upside-down corgis.
+"""
+
+# ╔═╡ aa541288-74aa-11eb-1edc-ab6d7786f271
+@bind α Slider(0:.01:1 , show_value=true, default = 1.0)
+
+# ╔═╡ c9dcac48-74aa-11eb-31a6-23357180c1c8
+α .* corgis .+ (1-α) .* upsidedown_corgis
+
+# ╔═╡ 30b1c1f0-7504-11eb-1be7-a9463caea809
+md"""
+The moment I did this with α = .5, I noticed my brain's tendency to see the 
+rightsisde-up corgis even though both have equal weight.  For me maybe
+around α = .39 which gives weight .61 to the upside-down corgis "feels" balanced
+to me.  I think this is what the field of psychology called psychometrics 
+tries to measure -- perhaps someone can tell me if there are studies of the
+brain's tendency to use world experience to prefer rightside-up corgis,
+and in particular to put a numerical value to this tendency.
+"""
+
+# ╔═╡ 1fe70e38-751b-11eb-25b8-c741e1726613
+md"""
+10 seconds with google and I found there is a thing about faces:
+[The Face Inversion effect](https://en.wikipedia.org/wiki/Face_inversion_effect#:~:text=The%20face%20inversion%20effect%20is,same%20for%20non%2Dfacial%20objects.&text=The%20most%20supported%20explanation%20for,is%20the%20configural%20information%20hypothesis)
+and also the [Thatcher Effect](https://en.wikipedia.org/wiki/Thatcher_effect#:~:text=The%20Thatcher%20effect%20or%20Thatcher,obvious%20in%20an%20upright%20face) seems related.
+
+... the article suggests objects don't suffer in the same way as faces,
+so I put forth that the phenomenon applies to corgi faces as much as human faces,
+suggesting maybe that corgi faces are processed in the **face processing** part of the brain,not the **object processing** part of the brain.
+
+Corgis are human, after all, right?
+
+(Note, this is 5 minutes of armchair science, not a professional opinion.)
+"""
+
+# ╔═╡ 215291ec-74a2-11eb-3476-0dab43fd5a5e
+md"""
+## 2.3 Fun with Photoshop (What does "filter" mean in this context?)
+"""
+
+# ╔═╡ 61db42c6-7505-11eb-1ddf-05e906234572
+md"""
+[Photshop Filter Reference](https://helpx.adobe.com/photoshop/using/filter-effects-reference.html)
+"""
+
+# ╔═╡ cdd4cffc-74b1-11eb-1aa4-e333cb8601d1
+md"""
+Let's play with photoshop if for no other reason, let's see what image transformations are available considered useful by the pros.
+
+Some worth emphasizing are
+1. Blur
+2. Sharpen
+3. Stylize -> Find Edges
+3. Pixelate
+4. Distort
+
+Some of these transformations (e.g. Blur, Sharpen, Find Edges) are examples of 
+convolutions which are very efficient, and show up these days in 
+machine learning particularly in image recognition.
+"""
+
+# ╔═╡ 7489a570-74a3-11eb-1d0b-09d41604ffe1
+md"""
+## 2.4 Image Filtering (convolutions)
+"""
+
+# ╔═╡ 8a8e3f5e-74b2-11eb-3eed-e5468e573e45
+md"""
+Last semester Grant Sanderson (3Blue1Brown) lectured in this course.  This lecture on convolutions in image processing was popular.  Let's watch an excerpt (from 1:04 to 2:48).  (We pick a few exercepts, but we wouldn't blame you if you just wanted to
+watch the whole video.)
+"""
+
+# ╔═╡ 5864294a-74a5-11eb-23ef-f38a582f2c2d
+html"""
+<div notthestyle="position: relative; right: 0; top: 0; z-index: 300;"><iframe src="https://www.youtube.com/embed/8rrHTtUzyZA?start=64&end=168" width=400 height=250  frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></div>
+"""
+
+# ╔═╡ fa9c465e-74b2-11eb-2f3c-4be0e7f93bb5
+md"""
+### Definition of convolutions and kernels
+"""
+
+# ╔═╡ 4fab4616-74b0-11eb-0088-6b50237d7d54
+md"""
+[Wikipedia Page on Kernels]
+(https://en.wikipedia.org/wiki/Kernel_(image_processing)#Details)
+"""
+
+# ╔═╡ 275bf7ac-74b3-11eb-32c3-cda1e4f1f8c2
+md"""
+### Computer Science: Complexity
+
+The number of multiplications = (Number of Pixels in the Image) * (Number of Cells in the kernel)
+"""
+
+# ╔═╡ 537c54e4-74b3-11eb-341f-951b4a1e0b40
+md"""
+Thought Problem: Why are small kernels better than large kernels from a complexity viewpoint?
+"""
+
+# ╔═╡ c6e340ee-751e-11eb-3ca7-69595b3693b7
+md"""
+### Computer Science: Architectures: GPUs or Graphical Processing Units
+
+Some important computations can be greatly accelerated through the use of specialized hardware such as the GPU processors that were originally designed as image renderers but it has turned out that these processors can be quite fast at other very regular computations.  Convolutions is a very GPU friendly operation due to its regular structure.
+"""
+
+# ╔═╡ 54448d18-7528-11eb-209a-9717affa0d02
+ kernelize(M) = OffsetArray( M, -1:1, -1:1)	       
+
+# ╔═╡ acbc563a-7528-11eb-3c38-75a5b66c9241
 begin
-	struct Paths
-	    m::Int
-	    n::Int
-	end
+	identity = [0 0 0 ; 0 1 0 ; 0 0 0]
+	edge_detect = [0 -1 0; -1 4 -1; 0 -1 0] 
+	sharpen = identity .+ edge_detect  # Superposition!
+	box_blur = [1 1 1;1 1 1;1 1 1]/9
+	∇x = [-1 0 1;-1 0 1;-1 0 1]/2 # centered deriv in x
+	∇y = ∇x'
 	
-	Base.iterate(p::Paths) = fill(1,p.m), fill(1,p.m) #start the iteration with 1's
-	
-	Base.IteratorSize(::Type{Paths}) = SizeUnknown()
-	
-	function Base.iterate(p::Paths, state)
-		if state ≠ fill(p.n,p.m) # end when each row has an n
-	      newstate = next(state,p.n)
-	      return newstate, newstate
-	    end
-	end
-	
-	
-	function next(path,n)
-	    k = length(path)
-		# start from the end and find the first element that can be updated by adding 1
-	    while  k≥2 && ( path[k]==n || path[k]+1 > path[k-1]+1 )
-	        k -= 1
-	    end   
-	    path[k] +=1 #add the one then reset the following elements
-	    for j = k+1 : length(path)
-	        path[j] = max(path[j-1]-1,1)
-	    end
-	    return(path)
-	end
-	
-
-	
-	function allpaths(m,n)
-     v=Vector{Int}[]
-	 paths = Paths(m,n)
-     for p ∈ paths
-        push!(v,copy(p))
-    end
-    v
-	end
+	kernels = [identity, edge_detect, sharpen, box_blur, ∇x, ∇y]
+	kernel_keys =["identity", "edge_detect", "sharpen", "box_blur", "∇x", "∇y"]
+	selections = kernel_keys .=> kernel_keys
+	kernel_matrix = Dict(kernel_keys .=> kernels)
+	md"$(@bind kernel_name Select(selections))"
 end
 
-# ╔═╡ d1c851ee-80d5-11eb-1ce4-357dfb1e638e
-begin
-	paths = allpaths(n,n)
-	numpaths = length(paths)
-	md"There are $numpaths paths to check."
-end
+# ╔═╡ 995392ee-752a-11eb-3394-0de331e24f40
+kernel_matrix[kernel_name]
 
-# ╔═╡ 5dd22d0e-80d6-11eb-0541-d77668309f6c
+# ╔═╡ d22903d6-7529-11eb-2dcd-132cd27104c2
+[imfilter( corgis, kernelize(kernel_matrix[kernel_name])) Gray.(1.5 .* abs.(imfilter( corgis, kernelize(kernel_matrix[kernel_name])))) ]
+
+# ╔═╡ 844ed844-74b3-11eb-2ee1-2de664b26bc6
 md"""
-Path $( @bind whichpath Slider(1:numpaths, show_value=true) )
+  ### Gaussian Filter
 """
 
-# ╔═╡ 84bb1f5c-80e5-11eb-0e55-83068948870c
-begin
-	fixedpaths = [p for p∈paths  if p[fixi]==fixj]
-	number_of_fixedpaths = length(fixedpaths)
-	md"Number of fixed paths = $number_of_fixedpaths"
-end
-
-# ╔═╡ ee2d787c-80e5-11eb-1930-0fcbe253643f
-@bind whichfixedpath Slider(1:number_of_fixedpaths)
-
-# ╔═╡ e5367534-80e5-11eb-341d-7b3e6ca4f111
-begin
-	
-	path = fixedpaths[whichfixedpath]
-	values = [ M[i,path[i]] for i=1:n]
-	nv = length(values)
-	thetitle = join([" $(values[i]) +" for i=1:nv-1 ]) * " $(values[end]) = $(sum(values))";
-	
-	rectangle(w, h, x, y) = Shape(x .+ [0,w,w,0], y .+ [0,0,h,h])
-
-	plot()
-	for i=1:n, j=1:n
-	   plot!(rectangle(1,1,i,j), opacity=.2, color=[:red,:white][1+rem(i+j,2) ])
-	   
-	end
-	for i=1:n, j=1:n
-	  annotate!((j+.5),n+2-(i+.5), M[i,j])
-	end
-	
-	annotate!((fixj+.5),n+2-(fixi+.5), M[fixi,fixj], :red)
-	
-	
-	for i = 1:n-1
-		i ≥ 	fixi ? c=:blue : c=:black
-		plot!([ path[i+1]+.5, path[i]+.5  ],[n-i+.5, n-i+1.5], color=c,  linewidth=4)
-	end
-	
-  xlabel!("")
-
-	
-	for i=1:n,j=1:n
-		plot!(rectangle(.4,.4,i+.3,j+.3), opacity=1, color=RGB(0,1,0), linewidth=0,fillcolor=[RGBA(1,.85,.85,.2),:white][1+rem(i+j,2)])
-	end
-	plot!(title=thetitle)
-	plot!(legend=false, aspectratio=1, xlims=(1,n+1), ylims=(1,n+1), axis=nothing)
-end
-
-# ╔═╡ 4e8c8052-8102-11eb-3e9f-01494b525ba0
+# ╔═╡ 4ffe927c-74b4-11eb-23a7-a18d7e51c75b
 md"""
-### Summing Paths Demo
+In our next Grant Sanderson segment from Fall 2020 (4:35 to 7:00), we hear about
+convolving images with a Gaussian kernel.  
 """
 
-# ╔═╡ bfa04a82-80d8-11eb-277a-f74429b09870
-begin
-	winnernum = argmin([sum( M[i,p[i]] for i=1:n) for p∈paths])
-	winner = paths[winnernum]
-	winnertotal = sum( M[i,winner[i]] for i=1:n);
-end
-
-# ╔═╡ 7191b674-80dc-11eb-24b3-518de83f465a
-md"""
-Our goal is to add the numbers on a path and find the minimal path.
-The winner is number $winnernum.
+# ╔═╡ 91109e5c-74b3-11eb-1f31-c50e436bc6e0
+html"""
+<div notthestyle="position: relative; right: 0; top: 0; z-index: 300;"><iframe src="https://www.youtube.com/embed/8rrHTtUzyZA?start=275&end=420" width=400 height=250  frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></div>
 """
 
-# ╔═╡ a7245c08-803f-11eb-0da9-2bed09872035
-let
-	
-	path = paths[whichpath]
-	values = [ M[i,path[i]] for i=1:n]
-	nv = length(values)
-	thetitle = join([" $(values[i]) +" for i=1:nv-1 ]) * " $(values[end]) = $(sum(values))";
-	
-	
-	rectangle(w, h, x, y) = Shape(x .+ [0,w,w,0], y .+ [0,0,h,h])
-	plot()
-	for i=1:n, j=1:n
-	   plot!(rectangle(1,1,i,j), opacity=.2, color=[:red,:white][1+rem(i+j,2) ])
-	   
-	end
-	for i=1:n, j=1:n
-	  annotate!((j+.5),n+2-(i+.5), M[i,j])
-	end
-	
-	# The winning path 
-		for i = 1:n-1
-		plot!([ winner[i+1]+.5, winner[i]+.5  ],[n-i+.5, n-i+1.5], color=RGB(1,.6,.6),  linewidth=4)
-	end
-	
-	
-	for i = 1:n-1
-		plot!([ path[i+1]+.5, path[i]+.5  ],[n-i+.5, n-i+1.5], color=:black,  linewidth=4)
-	end
-	
-	plot!(xlabel="winner total = $winnertotal", xguidefontcolor=RGB(1,.5,.5))
+# ╔═╡ 34109062-7525-11eb-10b3-d59d3a6dfda6
+round.(Kernel.gaussian(1), digits=3)
 
-	
-	for i=1:n,j=1:n
-		plot!(rectangle(.4,.4,i+.3,j+.3), opacity=1, color=RGB(0,1,0), linewidth=0,fillcolor=[RGBA(1,.85,.85,.2),:white][1+rem(i+j,2)])
-	end
-	plot!(title=thetitle)
-	plot!(legend=false, aspectratio=1, xlims=(1,n+1), ylims=(1,n+1), axis=nothing)
+# ╔═╡ 9ab89a3a-7525-11eb-186d-29e4b61deb7f
+md"""
+We could have defined this ourselves with calls to the exponential function.
+"""
+
+# ╔═╡ 50034058-7525-11eb-345b-3334e71ac50e
+begin
+	G = [exp( -(i^2+j^2)/2) for i=-2:2, j=-2:2]
+	round.(G ./ sum(G), digits=3)
 end
+
+# ╔═╡ c0aec7ae-7505-11eb-2822-a151aad48fc9
+md"""
+This is often known as Gausssian blur to emphasize the result of this operation.
+[Adobe on Gaussian blur](https://www.adobe.com/creativecloud/photography/discover/gaussian-blur.html).
+"""
+
+# ╔═╡ 628aea22-7521-11eb-2edc-39ac62683aea
+md"""
+Focus around 5:23
+"""
+
+# ╔═╡ 99eeb11c-7524-11eb-2154-df7d84976445
+@bind gparam Slider(0:9, show_value=true, default=1)
+
+# ╔═╡ 2ddcfb90-7520-11eb-2df7-172d07118b7e
+kernel = Kernel.gaussian(gparam)
+
+# ╔═╡ d62496b0-7524-11eb-3410-7177e7c7f8eb
+plotly()
+
+# ╔═╡ 6aa8a76e-7524-11eb-22b5-015aab4191b0
+surface([kernel;])
+
+# ╔═╡ ee93eeb2-7524-11eb-342d-0343d8aebf59
+md"""
+Note: black lines are contours
+"""
+
+# ╔═╡ 662d73b6-74b3-11eb-333d-f1323a001000
+md"""
+### Computer Science: Data Structure: Offset Arrays
+"""
+
+# ╔═╡ d127303a-7521-11eb-3507-7341a416211f
+kernel[0,0]
+
+# ╔═╡ d4581b56-7522-11eb-2c15-991c0c790e67
+kernel[-2,2]
+
+# ╔═╡ 40c15c3a-7523-11eb-1f2a-bd90b127dad2
+M = [ 1  2  3  4  5
+	  6  7  8  9 10
+	 11 12 13 14 15]
+
+# ╔═╡ 08642690-7523-11eb-00dd-63d4cf6513dc
+Z = OffsetArray(M, -1:1, -2:2)
+
+# ╔═╡ deac4cf2-7523-11eb-2832-7b9d31389b08
+the_indices = [ c.I for c ∈ CartesianIndices(Z)]
+
+# ╔═╡ 32887dfa-7524-11eb-35cd-051eff594fa9
+Z[1,-2]
+
+# ╔═╡ 0f765670-7506-11eb-2a37-931b15bb387f
+md"""
+## 2.5. Discrete vs Continuous
+"""
+
+# ╔═╡ 82737d28-7507-11eb-1e39-c7dc12e18882
+md"""
+Some folks only like discrete objects, others continuous.  The computer makes clear what many mathematicians already know, that while different language has evolved to describe discrete objects vs continuous objects, often the underlying conceptual idea is similar or the same.  Here is one analogy:
+"""
+
+# ╔═╡ 40d538b2-7506-11eb-116b-efeb16b3478d
+md"""
+### Blurring Kernels :: Integrals  ≡ Sharpening Kernels :: Derivatives
+"""
+
+# ╔═╡ df060a88-7507-11eb-034b-5346d67a0e0d
+md"""
+Think about integrals vs derivatives in one dimension.
+If you replace f(x) with g(x) = ∫ f(t) dt for x-r ≤ t ≤ x+r, that will blur or smooth out the features of f.  However if you take the derivative,you will emphasize the changes, i.e., you will sharpen or "edge-detect."
+"""
+
+# ╔═╡ 60c8db60-7506-11eb-1468-c989809c933a
+md"""
+## 2.6 Respect my Boundaries
+"""
+
+# ╔═╡ 8ed0be60-7506-11eb-2769-5f7da1c66243
+md"""
+Applying the convolution on a boundary requires special thought because it is literally an **edge case**.  Once again Grant said this so very well: (2:53-4:19)
+"""
+
+# ╔═╡ b9d636da-7506-11eb-37a6-3116d47b2787
+html"""
+<div notthestyle="position: relative; right: 0; top: 0; z-index: 300;"><iframe src="https://www.youtube.com/embed/8rrHTtUzyZA?start=173&end=259" width=400 height=250  frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe></div>
+"""
 
 # ╔═╡ 00000000-0000-0000-0000-000000000001
 PLUTO_PROJECT_TOML_CONTENTS = """
 [deps]
+ColorVectorSpace = "c3611d14-8923-5661-9e6a-0046d554d3a4"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
+FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
+ImageFiltering = "6a3955dd-da59-5b1f-98d4-e7296123deb5"
 ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
-Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
+ImageShow = "4e3cecfd-b093-5904-9786-8bbb286a6a31"
+OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PlutoUI = "7f904dfe-b85e-4ff6-b463-dae2292396a8"
+Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
 [compat]
+ColorVectorSpace = "~0.9.8"
 Colors = "~0.12.8"
+FileIO = "~1.14.0"
+ImageFiltering = "~0.7.1"
 ImageIO = "~0.6.2"
-Images = "~0.25.2"
+ImageShow = "~0.3.4"
+OffsetArrays = "~1.11.0"
 Plots = "~1.29.0"
 PlutoUI = "~0.7.38"
+Unitful = "~1.11.0"
 """
 
 # ╔═╡ 00000000-0000-0000-0000-000000000002
@@ -312,26 +516,8 @@ version = "3.3.3"
 [[ArgTools]]
 uuid = "0dad84c5-d112-42e6-8d28-ef12dabb789f"
 
-[[ArnoldiMethod]]
-deps = ["LinearAlgebra", "Random", "StaticArrays"]
-git-tree-sha1 = "62e51b39331de8911e4a7ff6f5aaf38a5f4cc0ae"
-uuid = "ec485272-7323-5ecc-a04f-4719b315124d"
-version = "0.2.0"
-
 [[Artifacts]]
 uuid = "56f22d72-fd6d-98f1-02f0-08ddc0907c33"
-
-[[AxisAlgorithms]]
-deps = ["LinearAlgebra", "Random", "SparseArrays", "WoodburyMatrices"]
-git-tree-sha1 = "66771c8d21c8ff5e3a93379480a2307ac36863f7"
-uuid = "13072b0f-2c55-5437-9ae7-d433b7a33950"
-version = "1.0.1"
-
-[[AxisArrays]]
-deps = ["Dates", "IntervalSets", "IterTools", "RangeArrays"]
-git-tree-sha1 = "cf6875678085aed97f52bfc493baaebeb6d40bcb"
-uuid = "39de3d68-74b9-583c-8d2d-e117c070f3a9"
-version = "0.4.5"
 
 [[Base64]]
 uuid = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
@@ -353,12 +539,6 @@ git-tree-sha1 = "4b859a208b2397a7a623a03449e4636bdb17bcf2"
 uuid = "83423d85-b0ee-5818-9007-b63ccbeb887a"
 version = "1.16.1+1"
 
-[[Calculus]]
-deps = ["LinearAlgebra"]
-git-tree-sha1 = "f641eb0a4f00c343bbc32346e1217b86f3ce9dad"
-uuid = "49dc2e85-a5d0-5ad3-a950-438e2897f1b9"
-version = "0.5.1"
-
 [[CatIndices]]
 deps = ["CustomUnitRanges", "OffsetArrays"]
 git-tree-sha1 = "a0f80a09780eed9b1d106a1bf62041c2efc995bc"
@@ -376,12 +556,6 @@ deps = ["ChainRulesCore", "LinearAlgebra", "Test"]
 git-tree-sha1 = "1e315e3f4b0b7ce40feded39c73049692126cf53"
 uuid = "9e997f8a-9a97-42d5-a9f1-ce6bfc15e2c0"
 version = "0.1.3"
-
-[[Clustering]]
-deps = ["Distances", "LinearAlgebra", "NearestNeighbors", "Printf", "SparseArrays", "Statistics", "StatsBase"]
-git-tree-sha1 = "75479b7df4167267d75294d14b58244695beb2ac"
-uuid = "aaaa29a8-35af-508c-8bc3-b662a17a0fe5"
-version = "0.14.2"
 
 [[ColorSchemes]]
 deps = ["ColorTypes", "ColorVectorSpace", "Colors", "FixedPointNumbers", "Random"]
@@ -422,17 +596,17 @@ git-tree-sha1 = "52cb3ec90e8a8bea0e62e275ba577ad0f74821f7"
 uuid = "ed09eef8-17a6-5b46-8889-db040fac31e3"
 version = "0.3.2"
 
+[[ConstructionBase]]
+deps = ["LinearAlgebra"]
+git-tree-sha1 = "f74e9d5388b8620b4cee35d4c5a618dd4dc547f4"
+uuid = "187b0558-2788-49d3-abe0-74a17ed4e7c9"
+version = "1.3.0"
+
 [[Contour]]
 deps = ["StaticArrays"]
 git-tree-sha1 = "9f02045d934dc030edad45944ea80dbd1f0ebea7"
 uuid = "d38c429a-6771-53c6-b99e-75d170b6e991"
 version = "0.5.7"
-
-[[CoordinateTransformations]]
-deps = ["LinearAlgebra", "StaticArrays"]
-git-tree-sha1 = "681ea870b918e7cff7111da58791d7f718067a19"
-uuid = "150eb455-5306-5404-9cee-2592286d6298"
-version = "0.6.2"
 
 [[CustomUnitRanges]]
 git-tree-sha1 = "1a3f97f907e6dd8983b744d2642651bb162a3f7a"
@@ -463,12 +637,6 @@ uuid = "ade2ca70-3891-5945-98fb-dc099432e06a"
 deps = ["Mmap"]
 uuid = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 
-[[Distances]]
-deps = ["LinearAlgebra", "SparseArrays", "Statistics", "StatsAPI"]
-git-tree-sha1 = "3258d0659f812acde79e8a74b11f17ac06d0ca04"
-uuid = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
-version = "0.10.7"
-
 [[Distributed]]
 deps = ["Random", "Serialization", "Sockets"]
 uuid = "8ba89e20-285c-5b6f-9357-94700520ee1b"
@@ -482,12 +650,6 @@ version = "0.8.6"
 [[Downloads]]
 deps = ["ArgTools", "LibCURL", "NetworkOptions"]
 uuid = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
-
-[[DualNumbers]]
-deps = ["Calculus", "NaNMath", "SpecialFunctions"]
-git-tree-sha1 = "5837a837389fccf076445fce071c8ddaea35a566"
-uuid = "fa6b7ba4-c1ee-5f82-b5fc-ecf0adba8f74"
-version = "0.6.8"
 
 [[EarCut_jll]]
 deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
@@ -597,12 +759,6 @@ git-tree-sha1 = "9b02998aba7bf074d14de89f9d37ca24a1a0b046"
 uuid = "78b55507-aeef-58d4-861c-77aaff3498b1"
 version = "0.21.0+0"
 
-[[Ghostscript_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "78e2c69783c9753a91cdae88a8d432be85a2ab5e"
-uuid = "61579ee1-b43e-5ca0-a5da-69d92c66a64b"
-version = "9.55.0+0"
-
 [[Glib_jll]]
 deps = ["Artifacts", "Gettext_jll", "JLLWrappers", "Libdl", "Libffi_jll", "Libiconv_jll", "Libmount_jll", "PCRE_jll", "Pkg", "Zlib_jll"]
 git-tree-sha1 = "a32d672ac2c967f3deb8a81d828afc739c838a06"
@@ -620,12 +776,6 @@ deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
 git-tree-sha1 = "344bf40dcab1073aca04aa0df4fb092f920e4011"
 uuid = "3b182d85-2403-5c21-9c21-1e1f0cc25472"
 version = "1.3.14+0"
-
-[[Graphs]]
-deps = ["ArnoldiMethod", "Compat", "DataStructures", "Distributed", "Inflate", "LinearAlgebra", "Random", "SharedArrays", "SimpleTraits", "SparseArrays", "Statistics"]
-git-tree-sha1 = "57c021de207e234108a6f1454003120a1bf350c4"
-uuid = "86223c79-3864-5bf0-83f7-82e725a168b6"
-version = "1.6.0"
 
 [[Grisu]]
 git-tree-sha1 = "53bb909d1151e57e2484c3d1b53e19552b887fb2"
@@ -662,35 +812,17 @@ git-tree-sha1 = "f7be53659ab06ddc986428d3a9dcc95f6fa6705a"
 uuid = "b5f81e59-6552-4d32-b1f0-c071b021bf89"
 version = "0.2.2"
 
-[[ImageAxes]]
-deps = ["AxisArrays", "ImageBase", "ImageCore", "Reexport", "SimpleTraits"]
-git-tree-sha1 = "c54b581a83008dc7f292e205f4c409ab5caa0f04"
-uuid = "2803e5a7-5153-5ecf-9a86-9b4c37f5f5ac"
-version = "0.6.10"
-
 [[ImageBase]]
 deps = ["ImageCore", "Reexport"]
 git-tree-sha1 = "b51bb8cae22c66d0f6357e3bcb6363145ef20835"
 uuid = "c817782e-172a-44cc-b673-b171935fbb9e"
 version = "0.1.5"
 
-[[ImageContrastAdjustment]]
-deps = ["ImageCore", "ImageTransformations", "Parameters"]
-git-tree-sha1 = "0d75cafa80cf22026cea21a8e6cf965295003edc"
-uuid = "f332f351-ec65-5f6a-b3d1-319c6670881a"
-version = "0.3.10"
-
 [[ImageCore]]
 deps = ["AbstractFFTs", "ColorVectorSpace", "Colors", "FixedPointNumbers", "Graphics", "MappedArrays", "MosaicViews", "OffsetArrays", "PaddedViews", "Reexport"]
 git-tree-sha1 = "9a5c62f231e5bba35695a20988fc7cd6de7eeb5a"
 uuid = "a09fc81d-aa75-5fe9-8630-4744c3626534"
 version = "0.9.3"
-
-[[ImageDistances]]
-deps = ["Distances", "ImageCore", "ImageMorphology", "LinearAlgebra", "Statistics"]
-git-tree-sha1 = "7a20463713d239a19cbad3f6991e404aca876bda"
-uuid = "51556ac3-7006-55f5-8cb3-34580c88182d"
-version = "0.2.15"
 
 [[ImageFiltering]]
 deps = ["CatIndices", "ComputationalResources", "DataStructures", "FFTViews", "FFTW", "ImageBase", "ImageCore", "LinearAlgebra", "OffsetArrays", "Reexport", "SparseArrays", "StaticArrays", "Statistics", "TiledIteration"]
@@ -704,59 +836,11 @@ git-tree-sha1 = "539682309e12265fbe75de8d83560c307af975bd"
 uuid = "82e4d734-157c-48bb-816b-45c225c6df19"
 version = "0.6.2"
 
-[[ImageMagick]]
-deps = ["FileIO", "ImageCore", "ImageMagick_jll", "InteractiveUtils", "Libdl", "Pkg", "Random"]
-git-tree-sha1 = "5bc1cb62e0c5f1005868358db0692c994c3a13c6"
-uuid = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
-version = "1.2.1"
-
-[[ImageMagick_jll]]
-deps = ["Artifacts", "Ghostscript_jll", "JLLWrappers", "JpegTurbo_jll", "Libdl", "Libtiff_jll", "Pkg", "Zlib_jll", "libpng_jll"]
-git-tree-sha1 = "f025b79883f361fa1bd80ad132773161d231fd9f"
-uuid = "c73af94c-d91f-53ed-93a7-00f77d67a9d7"
-version = "6.9.12+2"
-
-[[ImageMetadata]]
-deps = ["AxisArrays", "ImageAxes", "ImageBase", "ImageCore"]
-git-tree-sha1 = "36cbaebed194b292590cba2593da27b34763804a"
-uuid = "bc367c6b-8a6b-528e-b4bd-a4b897500b49"
-version = "0.9.8"
-
-[[ImageMorphology]]
-deps = ["ImageCore", "LinearAlgebra", "Requires", "TiledIteration"]
-git-tree-sha1 = "7668b123ecfd39a6ae3fc31c532b588999bdc166"
-uuid = "787d08f9-d448-5407-9aad-5290dd7ab264"
-version = "0.3.1"
-
-[[ImageQualityIndexes]]
-deps = ["ImageContrastAdjustment", "ImageCore", "ImageDistances", "ImageFiltering", "OffsetArrays", "Statistics"]
-git-tree-sha1 = "1d2d73b14198d10f7f12bf7f8481fd4b3ff5cd61"
-uuid = "2996bd0c-7a13-11e9-2da2-2f5ce47296a9"
-version = "0.3.0"
-
-[[ImageSegmentation]]
-deps = ["Clustering", "DataStructures", "Distances", "Graphs", "ImageCore", "ImageFiltering", "ImageMorphology", "LinearAlgebra", "MetaGraphs", "RegionTrees", "SimpleWeightedGraphs", "StaticArrays", "Statistics"]
-git-tree-sha1 = "36832067ea220818d105d718527d6ed02385bf22"
-uuid = "80713f31-8817-5129-9cf8-209ff8fb23e1"
-version = "1.7.0"
-
 [[ImageShow]]
 deps = ["Base64", "FileIO", "ImageBase", "ImageCore", "OffsetArrays", "StackViews"]
 git-tree-sha1 = "25f7784b067f699ae4e4cb820465c174f7022972"
 uuid = "4e3cecfd-b093-5904-9786-8bbb286a6a31"
 version = "0.3.4"
-
-[[ImageTransformations]]
-deps = ["AxisAlgorithms", "ColorVectorSpace", "CoordinateTransformations", "ImageBase", "ImageCore", "Interpolations", "OffsetArrays", "Rotations", "StaticArrays"]
-git-tree-sha1 = "42fe8de1fe1f80dab37a39d391b6301f7aeaa7b8"
-uuid = "02fcd773-0e25-5acc-982a-7f6622650795"
-version = "0.9.4"
-
-[[Images]]
-deps = ["Base64", "FileIO", "Graphics", "ImageAxes", "ImageBase", "ImageContrastAdjustment", "ImageCore", "ImageDistances", "ImageFiltering", "ImageIO", "ImageMagick", "ImageMetadata", "ImageMorphology", "ImageQualityIndexes", "ImageSegmentation", "ImageShow", "ImageTransformations", "IndirectArrays", "IntegralArrays", "Random", "Reexport", "SparseArrays", "StaticArrays", "Statistics", "StatsBase", "TiledIteration"]
-git-tree-sha1 = "03d1301b7ec885b266c0f816f338368c6c0b81bd"
-uuid = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
-version = "0.25.2"
 
 [[Imath_jll]]
 deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
@@ -779,12 +863,6 @@ git-tree-sha1 = "f550e6e32074c939295eb5ea6de31849ac2c9625"
 uuid = "83e8ac13-25f8-5344-8a64-a9f2b223428f"
 version = "0.5.1"
 
-[[IntegralArrays]]
-deps = ["ColorTypes", "FixedPointNumbers", "IntervalSets"]
-git-tree-sha1 = "509075560b9fce23fdb3ccb4cc97935f11a43aa0"
-uuid = "1d092043-8f09-5a30-832f-7509e371ab51"
-version = "0.1.4"
-
 [[IntelOpenMP_jll]]
 deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
 git-tree-sha1 = "d979e54b71da82f3a65b62553da4fc3d18c9004c"
@@ -794,18 +872,6 @@ version = "2018.0.3+2"
 [[InteractiveUtils]]
 deps = ["Markdown"]
 uuid = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
-
-[[Interpolations]]
-deps = ["AxisAlgorithms", "ChainRulesCore", "LinearAlgebra", "OffsetArrays", "Random", "Ratios", "Requires", "SharedArrays", "SparseArrays", "StaticArrays", "WoodburyMatrices"]
-git-tree-sha1 = "b7bc05649af456efc75d178846f47006c2c4c3c7"
-uuid = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
-version = "0.13.6"
-
-[[IntervalSets]]
-deps = ["Dates", "Statistics"]
-git-tree-sha1 = "eb381d885e30ef859068fce929371a8a5d06a914"
-uuid = "8197267c-284f-5f27-9208-e0e47529a953"
-version = "0.6.1"
 
 [[InverseFunctions]]
 deps = ["Test"]
@@ -827,12 +893,6 @@ version = "1.4.0"
 git-tree-sha1 = "a3f24677c21f5bbe9d2a714f95dcd58337fb2856"
 uuid = "82899510-4779-5014-852e-03e436cf321d"
 version = "1.0.0"
-
-[[JLD2]]
-deps = ["FileIO", "MacroTools", "Mmap", "OrderedCollections", "Pkg", "Printf", "Reexport", "TranscodingStreams", "UUIDs"]
-git-tree-sha1 = "81b9477b49402b47fbe7f7ae0b252077f53e4a08"
-uuid = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
-version = "0.4.22"
 
 [[JLLWrappers]]
 deps = ["Preferences"]
@@ -1007,12 +1067,6 @@ git-tree-sha1 = "e498ddeee6f9fdb4551ce855a46f54dbd900245f"
 uuid = "442fdcdd-2543-5da2-b0f3-8c86c306513e"
 version = "0.3.1"
 
-[[MetaGraphs]]
-deps = ["Graphs", "JLD2", "Random"]
-git-tree-sha1 = "2af69ff3c024d13bde52b34a2a7d6887d4e7b438"
-uuid = "626554b9-1ddb-594c-aa3c-2596fe9399a5"
-version = "0.7.1"
-
 [[Missings]]
 deps = ["DataAPI"]
 git-tree-sha1 = "bf210ce90b6c9eed32d25dbcae1ebc565df2687f"
@@ -1035,12 +1089,6 @@ uuid = "14a3606d-f60d-562e-9121-12d972cd8159"
 git-tree-sha1 = "b086b7ea07f8e38cf122f5016af580881ac914fe"
 uuid = "77ba4419-2d1f-58cd-9bb1-8ffee604a2e3"
 version = "0.3.7"
-
-[[NearestNeighbors]]
-deps = ["Distances", "StaticArrays"]
-git-tree-sha1 = "ded92de95031d4a8c61dfb6ba9adb6f1d8016ddd"
-uuid = "b8a86587-4115-5ab1-83bc-aa920d37bbce"
-version = "0.4.10"
 
 [[Netpbm]]
 deps = ["FileIO", "ImageCore"]
@@ -1124,12 +1172,6 @@ git-tree-sha1 = "03a7a85b76381a3d04c7a1656039197e70eda03d"
 uuid = "5432bcbf-9aad-5242-b902-cca2824c8663"
 version = "0.5.11"
 
-[[Parameters]]
-deps = ["OrderedCollections", "UnPack"]
-git-tree-sha1 = "34c0e9ad262e5f7fc75b10a9952ca7692cfc5fbe"
-uuid = "d96e819e-fc66-5662-9728-84c9c7592b0a"
-version = "0.12.3"
-
 [[Parsers]]
 deps = ["Dates"]
 git-tree-sha1 = "1285416549ccfcdf0c50d4997a94331e88d68413"
@@ -1204,12 +1246,6 @@ git-tree-sha1 = "c6c0f690d0cc7caddb74cef7aa847b824a16b256"
 uuid = "ea2cea3b-5b76-57ae-a6ef-0a8af62496e1"
 version = "5.15.3+1"
 
-[[Quaternions]]
-deps = ["DualNumbers", "LinearAlgebra", "Random"]
-git-tree-sha1 = "b327e4db3f2202a4efafe7569fcbe409106a1f75"
-uuid = "94ee1d12-ae83-5a48-8b1c-48b8ff168ae0"
-version = "0.5.6"
-
 [[REPL]]
 deps = ["InteractiveUtils", "Markdown", "Sockets", "Unicode"]
 uuid = "3fa0cd96-eef1-5676-8a61-b3b8758bbffb"
@@ -1217,17 +1253,6 @@ uuid = "3fa0cd96-eef1-5676-8a61-b3b8758bbffb"
 [[Random]]
 deps = ["SHA", "Serialization"]
 uuid = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-
-[[RangeArrays]]
-git-tree-sha1 = "b9039e93773ddcfc828f12aadf7115b4b4d225f5"
-uuid = "b3c3ace0-ae52-54e7-9d0b-2c1406fd6b9d"
-version = "0.3.2"
-
-[[Ratios]]
-deps = ["Requires"]
-git-tree-sha1 = "dc84268fe0e3335a62e315a3a7cf2afa7178a734"
-uuid = "c84ed2f1-dad5-54f0-aa8e-dbefe2724439"
-version = "0.4.3"
 
 [[RecipesBase]]
 git-tree-sha1 = "6bf3f380ff52ce0832ddd3a2a7b9538ed1bcca7d"
@@ -1245,12 +1270,6 @@ git-tree-sha1 = "45e428421666073eab6f2da5c9d310d99bb12f9b"
 uuid = "189a3867-3050-52da-a836-e630ba90ab69"
 version = "1.2.2"
 
-[[RegionTrees]]
-deps = ["IterTools", "LinearAlgebra", "StaticArrays"]
-git-tree-sha1 = "4618ed0da7a251c7f92e869ae1a19c74a7d2a7f9"
-uuid = "dee08c22-ab7f-5625-9660-a9af2021b33f"
-version = "0.3.2"
-
 [[RelocatableFolders]]
 deps = ["SHA", "Scratch"]
 git-tree-sha1 = "cdbd3b1338c72ce29d9584fdbe9e9b70eeb5adca"
@@ -1262,12 +1281,6 @@ deps = ["UUIDs"]
 git-tree-sha1 = "838a3a4188e2ded87a4f9f184b4b0d78a1e91cb7"
 uuid = "ae029012-a4dd-5104-9daa-d747884805df"
 version = "1.3.0"
-
-[[Rotations]]
-deps = ["LinearAlgebra", "Quaternions", "Random", "StaticArrays", "Statistics"]
-git-tree-sha1 = "3177100077c68060d63dd71aec209373c3ec339b"
-uuid = "6038ab10-8711-5258-84ad-4b1120ba62dc"
-version = "1.3.1"
 
 [[SHA]]
 uuid = "ea8e919c-243c-51af-8825-aaa63cd721ce"
@@ -1290,18 +1303,6 @@ deps = ["Dates", "Grisu"]
 git-tree-sha1 = "91eddf657aca81df9ae6ceb20b959ae5653ad1de"
 uuid = "992d4aef-0814-514b-bc4d-f2e9a6c4116f"
 version = "1.0.3"
-
-[[SimpleTraits]]
-deps = ["InteractiveUtils", "MacroTools"]
-git-tree-sha1 = "5d7e3f4e11935503d3ecaf7186eac40602e7d231"
-uuid = "699a6c99-e7fa-54fc-8d76-47d257e15c1d"
-version = "0.9.4"
-
-[[SimpleWeightedGraphs]]
-deps = ["Graphs", "LinearAlgebra", "Markdown", "SparseArrays", "Test"]
-git-tree-sha1 = "a6f404cc44d3d3b28c793ec0eb59af709d827e4e"
-uuid = "47aef6b3-ad0c-573a-a1e2-d07658019622"
-version = "1.2.1"
 
 [[Sixel]]
 deps = ["Dates", "FileIO", "ImageCore", "IndirectArrays", "OffsetArrays", "REPL", "libsixel_jll"]
@@ -1404,12 +1405,6 @@ git-tree-sha1 = "5683455224ba92ef59db72d10690690f4a8dc297"
 uuid = "06e1c1a7-607b-532d-9fad-de7d9aa2abac"
 version = "0.3.1"
 
-[[TranscodingStreams]]
-deps = ["Random", "Test"]
-git-tree-sha1 = "216b95ea110b5972db65aa90f88d8d89dcb8851c"
-uuid = "3bb67fe8-82b1-5028-8e26-92a6c54297fa"
-version = "0.9.6"
-
 [[Tricks]]
 git-tree-sha1 = "6bac775f2d42a611cdfcd1fb217ee719630c4175"
 uuid = "410a4b4d-49e4-4fbc-ab6d-cb71b17b3775"
@@ -1424,11 +1419,6 @@ version = "1.3.0"
 deps = ["Random", "SHA"]
 uuid = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
-[[UnPack]]
-git-tree-sha1 = "387c1f73762231e86e0c9c5443ce3b4a0a9a0c2b"
-uuid = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
-version = "1.0.2"
-
 [[Unicode]]
 uuid = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
 
@@ -1437,6 +1427,12 @@ deps = ["REPL"]
 git-tree-sha1 = "53915e50200959667e78a92a418594b428dffddf"
 uuid = "1cfade01-22cf-5700-b092-accc4b62d6e1"
 version = "0.4.1"
+
+[[Unitful]]
+deps = ["ConstructionBase", "Dates", "LinearAlgebra", "Random"]
+git-tree-sha1 = "b649200e887a487468b71821e2644382699f1b0f"
+uuid = "1986cc42-f94f-5a68-af5c-568840ba703d"
+version = "1.11.0"
 
 [[Unzip]]
 git-tree-sha1 = "34db80951901073501137bdbc3d5a8e7bbd06670"
@@ -1454,12 +1450,6 @@ deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
 git-tree-sha1 = "4528479aa01ee1b3b4cd0e6faef0e04cf16466da"
 uuid = "2381bf8a-dfd0-557d-9999-79630e7b1b91"
 version = "1.25.0+0"
-
-[[WoodburyMatrices]]
-deps = ["LinearAlgebra", "SparseArrays"]
-git-tree-sha1 = "de67fa59e33ad156a590055375a30b23c40299d3"
-uuid = "efce3f68-66dc-5838-9240-27a6d6f5f9b6"
-version = "0.5.5"
 
 [[XML2_jll]]
 deps = ["Artifacts", "JLLWrappers", "Libdl", "Libiconv_jll", "Pkg", "Zlib_jll"]
@@ -1671,30 +1661,91 @@ version = "0.9.1+5"
 """
 
 # ╔═╡ Cell order:
-# ╠═71b53b98-8038-11eb-0ea5-d953294e9f35
-# ╟─a84fdba4-80db-11eb-13dc-3f440653b2b9
-# ╟─938107f0-80ee-11eb-18cf-775802c43c2f
-# ╟─eb043a90-8102-11eb-3b78-d590a23c83f4
-# ╟─5994117c-8102-11eb-1b05-671b7cf87a7e
-# ╟─b4558306-804a-11eb-2719-5fd37c6fa281
-# ╟─bc631086-804a-11eb-216e-c955e2115f55
-# ╟─d1c851ee-80d5-11eb-1ce4-357dfb1e638e
-# ╟─7191b674-80dc-11eb-24b3-518de83f465a
-# ╟─5dd22d0e-80d6-11eb-0541-d77668309f6c
-# ╟─a7245c08-803f-11eb-0da9-2bed09872035
-# ╟─4e4d333e-8102-11eb-0ba1-0f0183d0d3c2
-# ╟─0f0e7456-8104-11eb-1d90-e9f0009e8789
-# ╟─4f969032-80e9-11eb-1ada-d1aa64960967
-# ╟─28f18aa2-8104-11eb-0c01-dbd14c760ecf
-# ╟─37ebfa3e-80e5-11eb-166c-4ff3471ab12d
-# ╟─84bb1f5c-80e5-11eb-0e55-83068948870c
-# ╟─ee2d787c-80e5-11eb-1930-0fcbe253643f
-# ╟─e5367534-80e5-11eb-341d-7b3e6ca4f111
-# ╟─4d81a6f4-8104-11eb-1f06-5bb7a56c8406
-# ╟─d9265982-80ed-11eb-3a5f-27712a23506b
-# ╟─ba4acb08-8104-11eb-1771-15bc5d8076fd
-# ╟─163bf8fe-80d0-11eb-2066-75439a533513
-# ╟─4e8c8052-8102-11eb-3e9f-01494b525ba0
-# ╠═bfa04a82-80d8-11eb-277a-f74429b09870
+# ╟─8d389d80-74a1-11eb-3452-f38eff03483b
+# ╟─9f1a72da-7532-11eb-079c-b7baccc6614a
+# ╠═86f770fe-74a1-11eb-01f7-5b3ecf057124
+# ╟─4d332c7e-74f8-11eb-1f49-a518246d1db8
+# ╟─f7689472-74a8-11eb-32a1-8379ae5c88e1
+# ╟─0f2f9004-74a8-11eb-01a2-973dbe80f166
+# ╠═962143a8-74a7-11eb-26c3-c10548f326ee
+# ╠═c2964c80-74f8-11eb-3a74-b1bdd9e4ae02
+# ╠═caf488d8-74f8-11eb-0075-0586d66c23c1
+# ╠═02dd4a02-74f9-11eb-3d1e-53d83cee8062
+# ╠═10ef13d2-74f9-11eb-2849-fb9f83db6ae9
+# ╠═b76a56f4-74a9-11eb-1739-fbfc5e4958e8
+# ╠═77fbf18a-74f9-11eb-1d9e-3f9d2097388f
+# ╠═bcb69db6-74f9-11eb-100a-29d1d23963ab
+# ╟─fc70c4d2-74f8-11eb-33f5-539c278ed6b6
+# ╟─2f7cde78-74a2-11eb-1e2f-81b5b2465819
+# ╟─e099815e-74a1-11eb-1541-033f6abe9f8e
+# ╟─e82a4dd8-74b0-11eb-1108-6b09e67a80c1
+# ╟─39552b7a-74fb-11eb-04e0-3981ada52c92
+# ╠═14f2b85e-74ad-11eb-2682-d9de646aedf3
+# ╠═516e73e2-74fb-11eb-213e-9dbd9472e0db
+# ╠═b5d0ef90-74fb-11eb-3126-792f954c7be7
+# ╠═754c3704-74fb-11eb-1199-2b9798d7251f
+# ╠═9eb917ba-74fb-11eb-0527-15e981ce9c6a
+# ╟─486d3022-74ff-11eb-1865-e15436bd9aad
+# ╟─b9da7332-74ff-11eb-241b-fb87e77d646a
+# ╟─339ccfca-74b1-11eb-0c35-774da6b189ed
+# ╟─8711c698-7500-11eb-2505-d35a4de169b4
+# ╟─84350cb8-7501-11eb-095e-8f1a7e015f25
+# ╠═91a1bca4-74aa-11eb-3917-1dfd73d0ad9c
+# ╠═8e698bdc-7501-11eb-1d2e-c336ccbde0b0
+# ╠═ab2bc924-7501-11eb-03ba-8dfc1ffe3f36
+# ╟─e11d6300-7501-11eb-239a-135596309d20
+# ╟─9a66c07e-7503-11eb-3127-7fce91b3a24a
+# ╟─47d40406-7502-11eb-2f43-cd5c848f25a6
+# ╠═9ce0b980-74aa-11eb-0678-01209451fb65
+# ╟─68821bf4-7502-11eb-0d3c-03d7a00fdba4
+# ╠═447e7c9e-74b1-11eb-27ea-71aa4338b11a
+# ╟─c9dff6f4-7503-11eb-2715-0bf9d3ece9e1
+# ╟─d834103c-7503-11eb-1a94-1fbad43801ff
+# ╠═aa541288-74aa-11eb-1edc-ab6d7786f271
+# ╠═c9dcac48-74aa-11eb-31a6-23357180c1c8
+# ╟─30b1c1f0-7504-11eb-1be7-a9463caea809
+# ╟─1fe70e38-751b-11eb-25b8-c741e1726613
+# ╟─215291ec-74a2-11eb-3476-0dab43fd5a5e
+# ╟─61db42c6-7505-11eb-1ddf-05e906234572
+# ╟─cdd4cffc-74b1-11eb-1aa4-e333cb8601d1
+# ╟─7489a570-74a3-11eb-1d0b-09d41604ffe1
+# ╟─8a8e3f5e-74b2-11eb-3eed-e5468e573e45
+# ╟─5864294a-74a5-11eb-23ef-f38a582f2c2d
+# ╟─fa9c465e-74b2-11eb-2f3c-4be0e7f93bb5
+# ╟─4fab4616-74b0-11eb-0088-6b50237d7d54
+# ╟─275bf7ac-74b3-11eb-32c3-cda1e4f1f8c2
+# ╟─537c54e4-74b3-11eb-341f-951b4a1e0b40
+# ╟─c6e340ee-751e-11eb-3ca7-69595b3693b7
+# ╠═54448d18-7528-11eb-209a-9717affa0d02
+# ╠═acbc563a-7528-11eb-3c38-75a5b66c9241
+# ╠═995392ee-752a-11eb-3394-0de331e24f40
+# ╠═d22903d6-7529-11eb-2dcd-132cd27104c2
+# ╟─844ed844-74b3-11eb-2ee1-2de664b26bc6
+# ╟─4ffe927c-74b4-11eb-23a7-a18d7e51c75b
+# ╟─91109e5c-74b3-11eb-1f31-c50e436bc6e0
+# ╠═34109062-7525-11eb-10b3-d59d3a6dfda6
+# ╟─9ab89a3a-7525-11eb-186d-29e4b61deb7f
+# ╠═50034058-7525-11eb-345b-3334e71ac50e
+# ╟─c0aec7ae-7505-11eb-2822-a151aad48fc9
+# ╟─628aea22-7521-11eb-2edc-39ac62683aea
+# ╠═99eeb11c-7524-11eb-2154-df7d84976445
+# ╠═2ddcfb90-7520-11eb-2df7-172d07118b7e
+# ╠═d62496b0-7524-11eb-3410-7177e7c7f8eb
+# ╠═6aa8a76e-7524-11eb-22b5-015aab4191b0
+# ╟─ee93eeb2-7524-11eb-342d-0343d8aebf59
+# ╟─662d73b6-74b3-11eb-333d-f1323a001000
+# ╠═d127303a-7521-11eb-3507-7341a416211f
+# ╠═d4581b56-7522-11eb-2c15-991c0c790e67
+# ╠═40c15c3a-7523-11eb-1f2a-bd90b127dad2
+# ╠═08642690-7523-11eb-00dd-63d4cf6513dc
+# ╠═deac4cf2-7523-11eb-2832-7b9d31389b08
+# ╠═32887dfa-7524-11eb-35cd-051eff594fa9
+# ╟─0f765670-7506-11eb-2a37-931b15bb387f
+# ╟─82737d28-7507-11eb-1e39-c7dc12e18882
+# ╟─40d538b2-7506-11eb-116b-efeb16b3478d
+# ╟─df060a88-7507-11eb-034b-5346d67a0e0d
+# ╟─60c8db60-7506-11eb-1468-c989809c933a
+# ╟─8ed0be60-7506-11eb-2769-5f7da1c66243
+# ╟─b9d636da-7506-11eb-37a6-3116d47b2787
 # ╟─00000000-0000-0000-0000-000000000001
 # ╟─00000000-0000-0000-0000-000000000002
